+++ examples/example_02.py
@@ -6,10 +6,10 @@ from redis.asyncio import Redis
 from redis_func_cache import LruTPolicy, RedisFuncCache
 
 # Create a redis client
-redis_client = Redis.from_url("redis://")
+redis_instance = Redis.from_url("redis://")
 
 # Create an lru cache, it connects Redis by previous created redis client
-lru_cache = RedisFuncCache(__name__, LruTPolicy(), redis_client)
+lru_cache = RedisFuncCache(__name__, LruTPolicy(), redis_instance=redis_instance)
 
 
 @lru_cache  # Decorate a function to cache its result